FEATURE: Special Schedules
At long last, it’s here.

to add a special(not fully implemented for api access), use redis-cli
and publish a message to the specials channel

example message:
{"date":"2015-12-18","schedule":[{"key_name":"4-030","name":"Block
1","shour":9,"smin":50,"ehour":10,"emin":55,"day":4}]}

The date is when the special should be used, in ISO8061 format. The
schedule is an array just like the regular database.